Split WASM wrapper runtime responsibilities
Extract executor, dispatch, polling, status, metadata, and shared-channel
responsibilities from `wrapper::mod` into focused sibling modules while
keeping the existing `WasmChannel` and `SharedWasmChannel` APIs reachable
through `crate::channels::wasm::wrapper`.

Keep the bindgen invocation and public re-exports in `mod.rs` so generated
WIT bindings remain available to the moved implementation blocks.
