+++ src/channels/wasm/wrapper/mod.rs
@@ -35,21 +35,16 @@ use std::time::Duration;
 use tokio::sync::{RwLock, mpsc, oneshot};
 use tokio_stream::wrappers::ReceiverStream;
 use uuid::Uuid;
-use wasmtime::Store;
-use wasmtime::component::{HasSelf, Linker};
 
 use crate::channels::wasm::capabilities::ChannelCapabilities;
 use crate::channels::wasm::error::WasmChannelError;
-use crate::channels::wasm::host::{
-    ChannelEmitRateLimiter, ChannelHostState, ChannelWorkspaceStore, EmittedMessage,
-};
+use crate::channels::wasm::host::{ChannelEmitRateLimiter, ChannelWorkspaceStore};
 use crate::channels::wasm::router::RegisteredEndpoint;
 use crate::channels::wasm::runtime::{PreparedChannelModule, WasmChannelRuntime};
 use crate::channels::wasm::schema::ChannelConfig;
 use crate::channels::{
     IncomingMessage, MessageStream, NativeChannel, OutgoingResponse, StatusUpdate,
 };
-use crate::db::{SettingKey, UserId};
 use crate::error::ChannelError;
 use crate::pairing::PairingStore;
 use crate::secrets::SecretsStore;
@@ -57,6 +52,12 @@ use crate::secrets::SecretsStore;
 mod attachments;
 mod convert;
 mod credentials;
+mod dispatch;
+mod executor;
+mod metadata;
+mod polling;
+mod shared;
+mod status;
 mod store;
 
 use attachments::read_attachments;
@@ -65,7 +66,10 @@ use convert::{
     wit_channel,
 };
 use credentials::resolve_channel_host_credentials;
+use metadata::do_update_broadcast_metadata;
 use store::{ChannelStoreData, ResolvedHostCredential};
+
+pub use shared::SharedWasmChannel;
 
 #[cfg(test)]
 use attachments::mime_from_extension;
@@ -148,36 +152,6 @@ pub struct WasmChannel {
     /// Secrets store for host-based credential injection.
     /// Used to pre-resolve credentials before each WASM callback.
     secrets_store: Option<Arc<dyn SecretsStore + Send + Sync>>,
-}
-
-/// Update broadcast metadata in memory and persist to the settings store when
-/// it changes. Extracted as a free function so both the `WasmChannel` instance
-/// method and the static polling helper share one implementation.
-async fn do_update_broadcast_metadata(
-    channel_name: &str,
-    metadata: &str,
-    last_broadcast_metadata: &tokio::sync::RwLock<Option<String>>,
-    settings_store: Option<&Arc<dyn crate::db::SettingsStore>>,
-) {
-    let mut guard = last_broadcast_metadata.write().await;
-    let changed = guard.as_deref() != Some(metadata);
-    *guard = Some(metadata.to_string());
-    drop(guard);
-
-    if changed && let Some(store) = settings_store {
-        let key = format!("channel_broadcast_metadata_{}", channel_name);
-        let value = serde_json::Value::String(metadata.to_string());
-        if let Err(e) = store
-            .set_setting(UserId::from("default"), SettingKey::from(key), &value)
-            .await
-        {
-            tracing::warn!(
-                channel = %channel_name,
-                "Failed to persist broadcast metadata: {}",
-                e
-            );
-        }
-    }
 }
 
 impl WasmChannel {
@@ -270,54 +244,6 @@ impl WasmChannel {
         &self.name
     }
 
-    /// Settings key for persisted broadcast metadata.
-    fn broadcast_metadata_key(&self) -> String {
-        format!("channel_broadcast_metadata_{}", self.name)
-    }
-
-    /// Update broadcast metadata in memory and persist if changed (best-effort).
-    ///
-    /// Compares with the current value to avoid redundant DB writes on every
-    /// incoming message (the chat_id rarely changes).
-    async fn update_broadcast_metadata(&self, metadata: &str) {
-        do_update_broadcast_metadata(
-            &self.name,
-            metadata,
-            &self.last_broadcast_metadata,
-            self.settings_store.as_ref(),
-        )
-        .await;
-    }
-
-    /// Load broadcast metadata from settings store on startup.
-    async fn load_broadcast_metadata(&self) {
-        if let Some(ref store) = self.settings_store {
-            match store
-                .get_setting(
-                    UserId::from("default"),
-                    SettingKey::from(self.broadcast_metadata_key()),
-                )
-                .await
-            {
-                Ok(Some(serde_json::Value::String(meta))) => {
-                    *self.last_broadcast_metadata.write().await = Some(meta);
-                    tracing::debug!(
-                        channel = %self.name,
-                        "Restored broadcast metadata from settings"
-                    );
-                }
-                Ok(_) => {}
-                Err(e) => {
-                    tracing::warn!(
-                        channel = %self.name,
-                        "Failed to load broadcast metadata: {}",
-                        e
-                    );
-                }
-            }
-        }
-    }
-
     /// Get the channel capabilities.
     pub fn capabilities(&self) -> &ChannelCapabilities {
         &self.capabilities
@@ -346,135 +272,6 @@ impl WasmChannel {
             });
         ws_cap.reader = Some(Arc::clone(store) as Arc<dyn crate::tools::wasm::WorkspaceReader>);
         caps
-    }
-
-    /// Add channel host functions to the linker using generated bindings.
-    ///
-    /// Uses the wasmtime::component::bindgen! generated `add_to_linker` function
-    /// to properly register all host functions with correct component model signatures.
-    fn add_host_functions(linker: &mut Linker<ChannelStoreData>) -> Result<(), WasmChannelError> {
-        // Add WASI support (required by the component adapter)
-        wasmtime_wasi::p2::add_to_linker_sync(linker).map_err(|e| {
-            WasmChannelError::Config(format!("Failed to add WASI functions: {}", e))
-        })?;
-
-        // Use the generated add_to_linker function from bindgen for our custom interface
-        near::agent::channel_host::add_to_linker::<_, HasSelf<_>>(linker, |state| state).map_err(
-            |e| WasmChannelError::Config(format!("Failed to add host functions: {}", e)),
-        )?;
-
-        Ok(())
-    }
-
-    /// Create a fresh store configured for WASM execution.
-    fn create_store(
-        runtime: &WasmChannelRuntime,
-        prepared: &PreparedChannelModule,
-        capabilities: &ChannelCapabilities,
-        credentials: HashMap<String, String>,
-        host_credentials: Vec<ResolvedHostCredential>,
-        pairing_store: Arc<PairingStore>,
-    ) -> Result<Store<ChannelStoreData>, WasmChannelError> {
-        let engine = runtime.engine();
-        let limits = &prepared.limits;
-
-        // Create fresh store with channel state (NEAR pattern: fresh instance per call)
-        let store_data = ChannelStoreData::new(
-            limits.memory_bytes,
-            &prepared.name,
-            capabilities.clone(),
-            credentials,
-            host_credentials,
-            pairing_store,
-        );
-        let mut store = Store::new(engine, store_data);
-
-        // Configure fuel if enabled
-        if runtime.config().fuel_config.enabled {
-            store
-                .set_fuel(limits.fuel)
-                .map_err(|e| WasmChannelError::Config(format!("Failed to set fuel: {}", e)))?;
-        }
-
-        // Configure epoch deadline for timeout backup
-        store.epoch_deadline_trap();
-        store.set_epoch_deadline(1);
-
-        // Set up resource limiter
-        store.limiter(|data| &mut data.limiter);
-
-        Ok(store)
-    }
-
-    /// Instantiate the WASM component using generated bindings.
-    fn instantiate_component(
-        runtime: &WasmChannelRuntime,
-        prepared: &PreparedChannelModule,
-        store: &mut Store<ChannelStoreData>,
-    ) -> Result<SandboxedChannel, WasmChannelError> {
-        let engine = runtime.engine();
-
-        // Use the pre-compiled component (no recompilation needed)
-        let component = prepared
-            .component()
-            .ok_or_else(|| {
-                WasmChannelError::Compilation("No compiled component available".to_string())
-            })?
-            .clone();
-
-        // Create linker and add host functions
-        let mut linker = Linker::new(engine);
-        Self::add_host_functions(&mut linker)?;
-
-        // Instantiate using the generated bindings
-        let instance = SandboxedChannel::instantiate(store, &component, &linker).map_err(|e| {
-            let msg = e.to_string();
-            if msg.contains("near:agent") || msg.contains("import") {
-                WasmChannelError::Instantiation(format!(
-                    "{msg}. This may indicate a WIT version mismatch — \
-                         the channel was compiled against a different WIT than the host supports \
-                         (host WIT: {}). Rebuild the channel against the current WIT.",
-                    crate::tools::wasm::WIT_CHANNEL_VERSION
-                ))
-            } else {
-                WasmChannelError::Instantiation(msg)
-            }
-        })?;
-
-        Ok(instance)
-    }
-
-    /// Map WASM execution errors to our error types.
-    fn map_wasm_error(e: wasmtime::Error, name: &str, fuel_limit: u64) -> WasmChannelError {
-        let error_str = e.to_string();
-        if error_str.contains("out of fuel") {
-            WasmChannelError::FuelExhausted {
-                name: name.to_string(),
-                limit: fuel_limit,
-            }
-        } else if error_str.contains("unreachable") {
-            WasmChannelError::Trapped {
-                name: name.to_string(),
-                reason: "unreachable code executed".to_string(),
-            }
-        } else {
-            WasmChannelError::Trapped {
-                name: name.to_string(),
-                reason: error_str,
-            }
-        }
-    }
-
-    /// Extract host state after callback execution.
-    fn extract_host_state(
-        store: &mut Store<ChannelStoreData>,
-        channel_name: &str,
-        capabilities: &ChannelCapabilities,
-    ) -> ChannelHostState {
-        std::mem::replace(
-            &mut store.data_mut().host_state,
-            ChannelHostState::new(channel_name, capabilities.clone()),
-        )
     }
 
     /// Execute the on_start callback.
@@ -1150,673 +947,6 @@ impl WasmChannel {
             }),
         }
     }
-
-    /// Execute a single on_status callback with a fresh WASM instance.
-    ///
-    /// Static method for use by the background typing repeat task (which
-    /// doesn't have access to `&self`).
-    #[allow(clippy::too_many_arguments)]
-    async fn execute_status(
-        channel_name: &str,
-        runtime: &Arc<WasmChannelRuntime>,
-        prepared: &Arc<PreparedChannelModule>,
-        capabilities: &ChannelCapabilities,
-        credentials: &RwLock<HashMap<String, String>>,
-        host_credentials: Vec<ResolvedHostCredential>,
-        pairing_store: Arc<PairingStore>,
-        timeout: Duration,
-        wit_update: wit_channel::StatusUpdate,
-    ) -> Result<(), WasmChannelError> {
-        if prepared.component().is_none() {
-            return Ok(());
-        }
-
-        let runtime = Arc::clone(runtime);
-        let prepared = Arc::clone(prepared);
-        let capabilities = capabilities.clone();
-        let credentials_snapshot = credentials.read().await.clone();
-        let channel_name_owned = channel_name.to_string();
-
-        let result = tokio::time::timeout(timeout, async move {
-            tokio::task::spawn_blocking(move || {
-                let mut store = Self::create_store(
-                    &runtime,
-                    &prepared,
-                    &capabilities,
-                    credentials_snapshot,
-                    host_credentials,
-                    pairing_store,
-                )?;
-                let instance = Self::instantiate_component(&runtime, &prepared, &mut store)?;
-
-                let channel_iface = instance.near_agent_channel();
-                channel_iface
-                    .call_on_status(&mut store, &wit_update)
-                    .map_err(|e| Self::map_wasm_error(e, &prepared.name, prepared.limits.fuel))?;
-
-                Ok(())
-            })
-            .await
-            .map_err(|e| WasmChannelError::ExecutionPanicked {
-                name: channel_name_owned.clone(),
-                reason: e.to_string(),
-            })?
-        })
-        .await;
-
-        match result {
-            Ok(Ok(())) => Ok(()),
-            Ok(Err(e)) => Err(e),
-            Err(_) => Err(WasmChannelError::Timeout {
-                name: channel_name.to_string(),
-                callback: "on_status".to_string(),
-            }),
-        }
-    }
-
-    /// Cancel the background typing indicator task if running.
-    async fn cancel_typing_task(&self) {
-        if let Some(handle) = self.typing_task.write().await.take() {
-            handle.abort();
-        }
-    }
-
-    /// Handle a status update, managing the typing repeat timer.
-    ///
-    /// On Thinking: fires on_status once, then spawns a background task
-    /// that repeats the call every 4 seconds (Telegram's typing indicator
-    /// expires after ~5s).
-    ///
-    /// On terminal or user-action-required states: cancels the repeat task,
-    /// then fires on_status once.
-    ///
-    /// On intermediate progress states (tool/auth/job/status updates), keeps
-    /// the typing repeater running and fires on_status once.
-    /// On StreamChunk: no-op (too noisy).
-    async fn handle_status_update(
-        &self,
-        status: StatusUpdate,
-        metadata: &serde_json::Value,
-    ) -> Result<(), ChannelError> {
-        fn is_terminal_text_status(msg: &str) -> bool {
-            let trimmed = msg.trim();
-            trimmed.eq_ignore_ascii_case("done")
-                || trimmed.eq_ignore_ascii_case("interrupted")
-                || trimmed.eq_ignore_ascii_case("awaiting approval")
-                || trimmed.eq_ignore_ascii_case("rejected")
-        }
-
-        match &status {
-            StatusUpdate::Thinking(_) => {
-                // Cancel any existing typing task
-                self.cancel_typing_task().await;
-
-                // Fire once immediately
-                if let Err(e) = self.call_on_status(&status, metadata).await {
-                    tracing::debug!(
-                        channel = %self.name,
-                        error = %e,
-                        "on_status(Thinking) failed (best-effort)"
-                    );
-                }
-
-                // Spawn background repeater
-                let channel_name = self.name.clone();
-                let runtime = Arc::clone(&self.runtime);
-                let prepared = Arc::clone(&self.prepared);
-                let capabilities = self.capabilities.clone();
-                let credentials = self.credentials.clone();
-                // Pre-resolve host credentials once for the lifetime of the repeater.
-                // Channels tokens rarely change, so a snapshot per-repeater is correct.
-                let repeater_host_credentials = resolve_channel_host_credentials(
-                    &self.capabilities,
-                    self.secrets_store.as_deref(),
-                )
-                .await;
-                let pairing_store = self.pairing_store.clone();
-                let callback_timeout = self.runtime.config().callback_timeout;
-                let wit_update = status_to_wit(&status, metadata);
-
-                let handle = tokio::spawn(async move {
-                    let mut interval = tokio::time::interval(Duration::from_secs(4));
-                    // Skip the first tick (we already fired above)
-                    interval.tick().await;
-
-                    loop {
-                        interval.tick().await;
-
-                        let wit_update_clone = clone_wit_status_update(&wit_update);
-                        let hc = repeater_host_credentials.clone();
-
-                        if let Err(e) = Self::execute_status(
-                            &channel_name,
-                            &runtime,
-                            &prepared,
-                            &capabilities,
-                            &credentials,
-                            hc,
-                            pairing_store.clone(),
-                            callback_timeout,
-                            wit_update_clone,
-                        )
-                        .await
-                        {
-                            tracing::debug!(
-                                channel = %channel_name,
-                                error = %e,
-                                "Typing repeat on_status failed (best-effort)"
-                            );
-                        }
-                    }
-                });
-
-                *self.typing_task.write().await = Some(handle);
-            }
-            StatusUpdate::StreamChunk(_) => {
-                // No-op, too noisy
-            }
-            StatusUpdate::ApprovalNeeded {
-                tool_name,
-                description,
-                parameters,
-                ..
-            } => {
-                // WASM channels (Telegram, Slack, etc.) cannot render
-                // interactive approval overlays.  Send the approval prompt
-                // as an actual message so the user can reply yes/no.
-                self.cancel_typing_task().await;
-
-                let params_preview = parameters
-                    .as_object()
-                    .map(|obj| {
-                        obj.iter()
-                            .map(|(k, v)| {
-                                let val = match v {
-                                    serde_json::Value::String(s) => {
-                                        if s.chars().count() > 80 {
-                                            let truncated: String = s.chars().take(77).collect();
-                                            format!("\"{}...\"", truncated)
-                                        } else {
-                                            format!("\"{}\"", s)
-                                        }
-                                    }
-                                    other => {
-                                        let s = other.to_string();
-                                        if s.chars().count() > 80 {
-                                            let truncated: String = s.chars().take(77).collect();
-                                            format!("{}...", truncated)
-                                        } else {
-                                            s
-                                        }
-                                    }
-                                };
-                                format!("  {}: {}", k, val)
-                            })
-                            .collect::<Vec<_>>()
-                            .join("\n")
-                    })
-                    .unwrap_or_default();
-
-                let prompt = format!(
-                    "Approval needed: {tool_name}\n\
-                     {description}\n\
-                     \n\
-                     Parameters:\n\
-                     {params_preview}\n\
-                     \n\
-                     Reply \"yes\" to approve, \"no\" to deny, or \"always\" to auto-approve."
-                );
-
-                let metadata_json = serde_json::to_string(metadata).unwrap_or_default();
-                if let Err(e) = self
-                    .call_on_respond(uuid::Uuid::new_v4(), &prompt, None, &metadata_json, &[])
-                    .await
-                {
-                    tracing::warn!(
-                        channel = %self.name,
-                        error = %e,
-                        "Failed to send approval prompt via on_respond, falling back to on_status"
-                    );
-                    // Fall back to status update (typing indicator)
-                    let _ = self.call_on_status(&status, metadata).await;
-                }
-            }
-            StatusUpdate::AuthRequired { .. } => {
-                // Waiting on user action: stop typing and fire once.
-                self.cancel_typing_task().await;
-
-                if let Err(e) = self.call_on_status(&status, metadata).await {
-                    tracing::debug!(
-                        channel = %self.name,
-                        error = %e,
-                        "on_status failed (best-effort)"
-                    );
-                }
-            }
-            StatusUpdate::Status(msg) if is_terminal_text_status(msg) => {
-                // Waiting on user or terminal states: stop typing and fire once.
-                self.cancel_typing_task().await;
-
-                if let Err(e) = self.call_on_status(&status, metadata).await {
-                    tracing::debug!(
-                        channel = %self.name,
-                        error = %e,
-                        "on_status failed (best-effort)"
-                    );
-                }
-            }
-            _ => {
-                // Intermediate progress status: keep any existing typing task alive.
-                if let Err(e) = self.call_on_status(&status, metadata).await {
-                    tracing::debug!(
-                        channel = %self.name,
-                        error = %e,
-                        "on_status failed (best-effort)"
-                    );
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Process emitted messages from a callback.
-    async fn process_emitted_messages(
-        &self,
-        messages: Vec<EmittedMessage>,
-    ) -> Result<(), WasmChannelError> {
-        tracing::info!(
-            channel = %self.name,
-            message_count = messages.len(),
-            "Processing emitted messages from WASM callback"
-        );
-
-        if messages.is_empty() {
-            tracing::debug!(channel = %self.name, "No messages emitted");
-            return Ok(());
-        }
-
-        let tx_guard = self.message_tx.read().await;
-        let Some(tx) = tx_guard.as_ref() else {
-            tracing::error!(
-                channel = %self.name,
-                count = messages.len(),
-                "Messages emitted but no sender available - channel may not be started!"
-            );
-            return Ok(());
-        };
-
-        let mut rate_limiter = self.rate_limiter.write().await;
-
-        for emitted in messages {
-            // Check rate limit
-            if !rate_limiter.check_and_record() {
-                tracing::warn!(
-                    channel = %self.name,
-                    "Message emission rate limited"
-                );
-                return Err(WasmChannelError::EmitRateLimited {
-                    name: self.name.clone(),
-                });
-            }
-
-            // Convert to IncomingMessage
-            let mut msg = IncomingMessage::new(&self.name, &emitted.user_id, &emitted.content);
-
-            if let Some(name) = emitted.user_name {
-                msg = msg.with_user_name(name);
-            }
-
-            if let Some(thread_id) = emitted.thread_id {
-                msg = msg.with_thread(thread_id);
-            }
-
-            // Convert attachments
-            if !emitted.attachments.is_empty() {
-                let incoming_attachments = emitted
-                    .attachments
-                    .iter()
-                    .map(|a| crate::channels::IncomingAttachment {
-                        id: a.id.clone(),
-                        kind: crate::channels::AttachmentKind::from_mime_type(&a.mime_type),
-                        mime_type: a.mime_type.clone(),
-                        filename: a.filename.clone(),
-                        size_bytes: a.size_bytes,
-                        source_url: a.source_url.clone(),
-                        storage_key: a.storage_key.clone(),
-                        extracted_text: a.extracted_text.clone(),
-                        data: a.data.clone(),
-                        duration_secs: a.duration_secs,
-                    })
-                    .collect();
-                msg = msg.with_attachments(incoming_attachments);
-            }
-
-            // Parse metadata JSON
-            if let Ok(metadata) = serde_json::from_str(&emitted.metadata_json) {
-                msg = msg.with_metadata(metadata);
-                // Store for broadcast routing (chat_id etc.)
-                self.update_broadcast_metadata(&emitted.metadata_json).await;
-            }
-
-            // Send to stream
-            tracing::info!(
-                channel = %self.name,
-                user_id = %emitted.user_id,
-                content_len = emitted.content.len(),
-                attachment_count = msg.attachments.len(),
-                "Sending emitted message to agent"
-            );
-
-            if tx.send(msg).await.is_err() {
-                tracing::error!(
-                    channel = %self.name,
-                    "Failed to send emitted message, channel closed"
-                );
-                break;
-            }
-
-            tracing::info!(
-                channel = %self.name,
-                "Message successfully sent to agent queue"
-            );
-        }
-
-        Ok(())
-    }
-
-    /// Start the polling loop if configured.
-    ///
-    /// Since we can't hold `Arc<Self>` from `&self`, we pass all the components
-    /// needed for polling to a spawned task. Each poll tick creates a fresh WASM
-    /// instance (matching our "fresh instance per callback" pattern).
-    fn start_polling(&self, interval: Duration, shutdown_rx: oneshot::Receiver<()>) {
-        let channel_name = self.name.clone();
-        let runtime = Arc::clone(&self.runtime);
-        let prepared = Arc::clone(&self.prepared);
-        let poll_capabilities = self.capabilities.clone();
-        let capabilities = Self::inject_workspace_reader(&self.capabilities, &self.workspace_store);
-        let message_tx = self.message_tx.clone();
-        let rate_limiter = self.rate_limiter.clone();
-        let credentials = self.credentials.clone();
-        let pairing_store = self.pairing_store.clone();
-        let callback_timeout = self.runtime.config().callback_timeout;
-        let workspace_store = self.workspace_store.clone();
-        let last_broadcast_metadata = self.last_broadcast_metadata.clone();
-        let settings_store = self.settings_store.clone();
-        let poll_secrets_store = self.secrets_store.clone();
-
-        tokio::spawn(async move {
-            let mut interval_timer = tokio::time::interval(interval);
-            let mut shutdown = std::pin::pin!(shutdown_rx);
-
-            loop {
-                tokio::select! {
-                    _ = interval_timer.tick() => {
-                        tracing::debug!(
-                            channel = %channel_name,
-                            "Polling tick - calling on_poll"
-                        );
-
-                        // Pre-resolve host credentials for this tick
-                        let host_credentials = resolve_channel_host_credentials(
-                            &poll_capabilities,
-                            poll_secrets_store.as_deref(),
-                        )
-                        .await;
-
-                        // Execute on_poll with fresh WASM instance
-                        let result = Self::execute_poll(
-                            &channel_name,
-                            &runtime,
-                            &prepared,
-                            &capabilities,
-                            &credentials,
-                            host_credentials,
-                            pairing_store.clone(),
-                            callback_timeout,
-                            &workspace_store,
-                        ).await;
-
-                        match result {
-                            Ok(emitted_messages) => {
-                                // Process any emitted messages
-                                if !emitted_messages.is_empty()
-                                    && let Err(e) = Self::dispatch_emitted_messages(
-                                        &channel_name,
-                                        emitted_messages,
-                                        &message_tx,
-                                        &rate_limiter,
-                                        &last_broadcast_metadata,
-                                        settings_store.as_ref(),
-                                    ).await {
-                                        tracing::warn!(
-                                            channel = %channel_name,
-                                            error = %e,
-                                            "Failed to dispatch emitted messages from poll"
-                                        );
-                                    }
-                            }
-                            Err(e) => {
-                                tracing::warn!(
-                                    channel = %channel_name,
-                                    error = %e,
-                                    "Polling callback failed"
-                                );
-                            }
-                        }
-                    }
-                    _ = &mut shutdown => {
-                        tracing::info!(
-                            channel = %channel_name,
-                            "Polling stopped"
-                        );
-                        break;
-                    }
-                }
-            }
-        });
-    }
-
-    /// Execute a single poll callback with a fresh WASM instance.
-    ///
-    /// Returns any emitted messages from the callback. Pending workspace writes
-    /// are committed to the shared `ChannelWorkspaceStore` so state persists
-    /// across poll ticks (e.g., Telegram polling offset).
-    #[allow(clippy::too_many_arguments)]
-    async fn execute_poll(
-        channel_name: &str,
-        runtime: &Arc<WasmChannelRuntime>,
-        prepared: &Arc<PreparedChannelModule>,
-        capabilities: &ChannelCapabilities,
-        credentials: &RwLock<HashMap<String, String>>,
-        host_credentials: Vec<ResolvedHostCredential>,
-        pairing_store: Arc<PairingStore>,
-        timeout: Duration,
-        workspace_store: &Arc<ChannelWorkspaceStore>,
-    ) -> Result<Vec<EmittedMessage>, WasmChannelError> {
-        // Skip if no WASM bytes (testing mode)
-        if prepared.component().is_none() {
-            tracing::debug!(
-                channel = %channel_name,
-                "WASM channel on_poll called (no WASM module)"
-            );
-            return Ok(Vec::new());
-        }
-
-        let runtime = Arc::clone(runtime);
-        let prepared = Arc::clone(prepared);
-        let capabilities = Self::inject_workspace_reader(capabilities, workspace_store);
-        let credentials_snapshot = credentials.read().await.clone();
-        let channel_name_owned = channel_name.to_string();
-        let workspace_store = Arc::clone(workspace_store);
-
-        // Execute in blocking task with timeout
-        let result = tokio::time::timeout(timeout, async move {
-            tokio::task::spawn_blocking(move || {
-                let mut store = Self::create_store(
-                    &runtime,
-                    &prepared,
-                    &capabilities,
-                    credentials_snapshot,
-                    host_credentials,
-                    pairing_store,
-                )?;
-                let instance = Self::instantiate_component(&runtime, &prepared, &mut store)?;
-
-                // Call on_poll using the generated typed interface
-                let channel_iface = instance.near_agent_channel();
-                channel_iface
-                    .call_on_poll(&mut store)
-                    .map_err(|e| Self::map_wasm_error(e, &prepared.name, prepared.limits.fuel))?;
-
-                let mut host_state =
-                    Self::extract_host_state(&mut store, &prepared.name, &capabilities);
-
-                // Commit pending workspace writes to the persistent store
-                let pending_writes = host_state.take_pending_writes();
-                workspace_store.commit_writes(&pending_writes);
-
-                Ok(host_state)
-            })
-            .await
-            .map_err(|e| WasmChannelError::ExecutionPanicked {
-                name: channel_name_owned.clone(),
-                reason: e.to_string(),
-            })?
-        })
-        .await;
-
-        match result {
-            Ok(Ok(mut host_state)) => {
-                let emitted = host_state.take_emitted_messages();
-                tracing::debug!(
-                    channel = %channel_name,
-                    emitted_count = emitted.len(),
-                    "WASM channel on_poll completed"
-                );
-                Ok(emitted)
-            }
-            Ok(Err(e)) => Err(e),
-            Err(_) => Err(WasmChannelError::Timeout {
-                name: channel_name.to_string(),
-                callback: "on_poll".to_string(),
-            }),
-        }
-    }
-
-    /// Dispatch emitted messages to the message channel.
-    ///
-    /// This is a static helper used by the polling loop since it doesn't have
-    /// access to `&self`.
-    async fn dispatch_emitted_messages(
-        channel_name: &str,
-        messages: Vec<EmittedMessage>,
-        message_tx: &RwLock<Option<mpsc::Sender<IncomingMessage>>>,
-        rate_limiter: &RwLock<ChannelEmitRateLimiter>,
-        last_broadcast_metadata: &tokio::sync::RwLock<Option<String>>,
-        settings_store: Option<&Arc<dyn crate::db::SettingsStore>>,
-    ) -> Result<(), WasmChannelError> {
-        tracing::info!(
-            channel = %channel_name,
-            message_count = messages.len(),
-            "Processing emitted messages from polling callback"
-        );
-
-        let tx_guard = message_tx.read().await;
-        let Some(tx) = tx_guard.as_ref() else {
-            tracing::error!(
-                channel = %channel_name,
-                count = messages.len(),
-                "Messages emitted but no sender available - channel may not be started!"
-            );
-            return Ok(());
-        };
-
-        let mut limiter = rate_limiter.write().await;
-
-        for emitted in messages {
-            // Check rate limit
-            if !limiter.check_and_record() {
-                tracing::warn!(
-                    channel = %channel_name,
-                    "Message emission rate limited"
-                );
-                return Err(WasmChannelError::EmitRateLimited {
-                    name: channel_name.to_string(),
-                });
-            }
-
-            // Convert to IncomingMessage
-            let mut msg = IncomingMessage::new(channel_name, &emitted.user_id, &emitted.content);
-
-            if let Some(name) = emitted.user_name {
-                msg = msg.with_user_name(name);
-            }
-
-            if let Some(thread_id) = emitted.thread_id {
-                msg = msg.with_thread(thread_id);
-            }
-
-            // Convert attachments
-            if !emitted.attachments.is_empty() {
-                let incoming_attachments = emitted
-                    .attachments
-                    .iter()
-                    .map(|a| crate::channels::IncomingAttachment {
-                        id: a.id.clone(),
-                        kind: crate::channels::AttachmentKind::from_mime_type(&a.mime_type),
-                        mime_type: a.mime_type.clone(),
-                        filename: a.filename.clone(),
-                        size_bytes: a.size_bytes,
-                        source_url: a.source_url.clone(),
-                        storage_key: a.storage_key.clone(),
-                        extracted_text: a.extracted_text.clone(),
-                        data: a.data.clone(),
-                        duration_secs: a.duration_secs,
-                    })
-                    .collect();
-                msg = msg.with_attachments(incoming_attachments);
-            }
-
-            // Parse metadata JSON
-            if let Ok(metadata) = serde_json::from_str(&emitted.metadata_json) {
-                msg = msg.with_metadata(metadata);
-                // Store for broadcast routing (chat_id etc.)
-                do_update_broadcast_metadata(
-                    channel_name,
-                    &emitted.metadata_json,
-                    last_broadcast_metadata,
-                    settings_store,
-                )
-                .await;
-            }
-
-            // Send to stream
-            tracing::info!(
-                channel = %channel_name,
-                user_id = %emitted.user_id,
-                content_len = emitted.content.len(),
-                attachment_count = msg.attachments.len(),
-                "Sending polled message to agent"
-            );
-
-            if tx.send(msg).await.is_err() {
-                tracing::error!(
-                    channel = %channel_name,
-                    "Failed to send polled message, channel closed"
-                );
-                break;
-            }
-
-            tracing::info!(
-                channel = %channel_name,
-                "Message successfully sent to agent queue"
-            );
-        }
-
-        Ok(())
-    }
 }
 
 impl NativeChannel for WasmChannel {
@@ -2004,81 +1134,6 @@ impl std::fmt::Debug for WasmChannel {
             .field("name", &self.name)
             .field("prepared", &self.prepared.name)
             .finish()
-    }
-}
-
-// ============================================================================
-// Shared Channel Wrapper
-// ============================================================================
-
-/// A wrapper around `Arc<WasmChannel>` that implements `Channel`.
-///
-/// This allows sharing the same WasmChannel instance between:
-/// - The WasmChannelRouter (for webhook handling)
-/// - The ChannelManager (for message streaming and responses)
-pub struct SharedWasmChannel {
-    inner: Arc<WasmChannel>,
-}
-
-impl SharedWasmChannel {
-    /// Create a new shared wrapper.
-    pub fn new(channel: Arc<WasmChannel>) -> Self {
-        Self { inner: channel }
-    }
-
-    /// Get the inner Arc.
-    pub fn inner(&self) -> &Arc<WasmChannel> {
-        &self.inner
-    }
-}
-
-impl std::fmt::Debug for SharedWasmChannel {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("SharedWasmChannel")
-            .field("inner", &self.inner)
-            .finish()
-    }
-}
-
-impl NativeChannel for SharedWasmChannel {
-    fn name(&self) -> &str {
-        self.inner.name()
-    }
-
-    async fn start(&self) -> Result<MessageStream, ChannelError> {
-        self.inner.start().await
-    }
-
-    async fn respond(
-        &self,
-        msg: &IncomingMessage,
-        response: OutgoingResponse,
-    ) -> Result<(), ChannelError> {
-        self.inner.respond(msg, response).await
-    }
-
-    async fn broadcast(
-        &self,
-        user_id: &str,
-        response: OutgoingResponse,
-    ) -> Result<(), ChannelError> {
-        self.inner.broadcast(user_id, response).await
-    }
-
-    async fn send_status(
-        &self,
-        status: StatusUpdate,
-        metadata: &serde_json::Value,
-    ) -> Result<(), ChannelError> {
-        self.inner.send_status(status, metadata).await
-    }
-
-    async fn health_check(&self) -> Result<(), ChannelError> {
-        self.inner.health_check().await
-    }
-
-    async fn shutdown(&self) -> Result<(), ChannelError> {
-        self.inner.shutdown().await
     }
 }
