+++ src/channels/wasm/wrapper/shared.rs
@@ -1,0 +1,85 @@
+//! Shared `Arc<WasmChannel>` wrapper for native channel trait use.
+
+use std::sync::Arc;
+
+use crate::channels::{
+    IncomingMessage, MessageStream, NativeChannel, OutgoingResponse, StatusUpdate,
+};
+use crate::error::ChannelError;
+
+use super::WasmChannel;
+
+// ============================================================================
+// Shared Channel Wrapper
+// ============================================================================
+
+/// A wrapper around `Arc<WasmChannel>` that implements `Channel`.
+///
+/// This allows sharing the same WasmChannel instance between:
+/// - The WasmChannelRouter (for webhook handling)
+/// - The ChannelManager (for message streaming and responses)
+pub struct SharedWasmChannel {
+    inner: Arc<WasmChannel>,
+}
+
+impl SharedWasmChannel {
+    /// Create a new shared wrapper.
+    pub fn new(channel: Arc<WasmChannel>) -> Self {
+        Self { inner: channel }
+    }
+
+    /// Get the inner Arc.
+    pub fn inner(&self) -> &Arc<WasmChannel> {
+        &self.inner
+    }
+}
+
+impl std::fmt::Debug for SharedWasmChannel {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("SharedWasmChannel")
+            .field("inner", &self.inner)
+            .finish()
+    }
+}
+
+impl NativeChannel for SharedWasmChannel {
+    fn name(&self) -> &str {
+        self.inner.name()
+    }
+
+    async fn start(&self) -> Result<MessageStream, ChannelError> {
+        self.inner.start().await
+    }
+
+    async fn respond(
+        &self,
+        msg: &IncomingMessage,
+        response: OutgoingResponse,
+    ) -> Result<(), ChannelError> {
+        self.inner.respond(msg, response).await
+    }
+
+    async fn broadcast(
+        &self,
+        user_id: &str,
+        response: OutgoingResponse,
+    ) -> Result<(), ChannelError> {
+        self.inner.broadcast(user_id, response).await
+    }
+
+    async fn send_status(
+        &self,
+        status: StatusUpdate,
+        metadata: &serde_json::Value,
+    ) -> Result<(), ChannelError> {
+        self.inner.send_status(status, metadata).await
+    }
+
+    async fn health_check(&self) -> Result<(), ChannelError> {
+        self.inner.health_check().await
+    }
+
+    async fn shutdown(&self) -> Result<(), ChannelError> {
+        self.inner.shutdown().await
+    }
+}
